00_02

+++ main.cpp
@@ -4,11 +4,12 @@ int WINAPI WinMain(HINSTANCE, HINSTANCE, LPSTR, int) {
 	OutputDebugStringA("Hello,Directx!\n");
 
 	int a = 4;
-	int b = 4.7f;
+	float b = 4.7f;
 
 	a = b;
 
 	//メインループ	9p
+
